requirements.txt
PySide6-Fluent-Widgets[full]
beautifulsoup4
mysql-connector-python
requests
sqlalchemy